requirements.txt
Django==4.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-uuidfield==0.5.0
drf-yasg==1.21.5
